api/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
numpy
temp
Pillow